backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
fitparse==1.2.0
fitdecode==0.11.0
gpxpy==1.6.2
aiofiles==23.2.1
pandas==2.2.0
httpx==0.27.2
cryptography==44.0.1
pytest==8.3.5
pytest-asyncio==0.25.3
